requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aenum=3.1.15=pypi_0
aiohttp=3.9.0=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.7.1=pypi_0
appnope=0.1.3=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blinker=1.7.0=pypi_0
bs4=0.0.1=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2023.08.22=hecd8cb5_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
comm=0.2.0=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.5=pypi_0
cycler=0.12.1=pypi_0
dask=2023.11.0=pypi_0
dataclasses-json=0.6.2=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
diskcache=5.6.3=pypi_0
distro=1.8.0=pypi_0
dnspython=2.4.2=pypi_0
eva-decord=0.6.1=pypi_0
evadb=0.3.9=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
faiss-cpu=1.7.4=pypi_0
fastjsonschema=2.19.0=pypi_0
filelock=3.13.1=pypi_0
flask=3.0.0=pypi_0
fonttools=4.45.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
gpt4all=2.0.2=pypi_0
greenlet=3.0.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.25.2=pypi_0
huggingface-hub=0.19.4=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.8.0=pypi_0
iniconfig=2.0.0=pypi_0
ipykernel=6.26.0=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.11.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.0=pypi_0
langchain=0.0.340=pypi_0
langsmith=0.0.66=pypi_0
lark=1.1.8=pypi_0
libffi=3.4.4=hecd8cb5_0
locket=1.0.0=pypi_0
loguru=0.7.2=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.6.8=pypi_0
nbformat=5.9.2=pypi_0
nbmake=1.4.6=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.8=pypi_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.2=pypi_0
openai=1.3.5=pypi_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openssl=3.0.12=hca72f7f_0
packaging=23.2=pypi_0
pandas=2.1.3=pypi_0
parso=0.8.3=pypi_0
partd=1.4.1=pypi_0
pdfminer-six=20221105=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.1.0=pypi_0
pinecone-client=2.2.4=pypi_0
pip=23.3.1=py310hecd8cb5_0
platformdirs=4.0.0=pypi_0
pluggy=1.3.0=pypi_0
prompt-toolkit=3.0.41=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py=1.11.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.13=pypi_0
pygments=2.17.2=pypi_0
pymupdf=1.22.5=pypi_0
pyparsing=3.1.1=pypi_0
pytest=7.4.3=pypi_0
python=3.10.13=h5ee71fb_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
rapidfuzz=3.5.2=pypi_0
readline=8.2=hca72f7f_0
referencing=0.31.0=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
retry=0.9.2=pypi_0
rpds-py=0.13.1=pypi_0
safetensors=0.4.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.13.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py310hecd8cb5_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.23=pypi_0
sqlalchemy-utils=0.41.1=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack-data=0.6.3=pypi_0
swifter=1.4.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
thefuzz=0.20.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.2.0=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.15.0=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pypi_0
transformers=4.35.2=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
ultralytics=8.0.217=pypi_0
urllib3=2.1.0=pypi_0
wcwidth=0.2.12=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310hecd8cb5_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.2=h6c40b1e_0
yarl=1.9.3=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h4dc903c_0
